[lib][web][native] Refactor actions in upload-actions.js
Summary:
issue: https://linear.app/comm/issue/ENG-4681/create-action-objects-for-all-actions
`uploadMultimedia` will be replaced by `useBlobServiceUpload` once blob service starts working. `uploadMediaMetadata` was not being used and is not needed anymore, as explained in https://linear.app/comm/issue/ENG-5174/get-rid-of-uploadmediametadata-action

Test Plan: Tested the `deleteUpload` action. The change to `blobServiceUpload` is the same as the changes to multiple actions before, so I think  it's fair to assume that it's working after countless other actions with the same changes have been tested. The problem with testing it is that it calls the blob service and would break the db state

Reviewers: michal, kamil, ginsu

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9399