[lib] Stop using Ethers.js client from Wagmi
Summary:
In Wagmi 1.0 they switch from Ethers.js to their in-house alternative Viem.

Our `ENSCache` code is built on Ethers.js. I'll consider migrating that to Viem, but for now the easiest thing is to decouple it.

Depends on D9432

Test Plan:
In combination with other diffs in the stack, I tested:
- Mobile sign in with Rainbow wallet using Rainbow login
- Mobile sign in with Rainbow wallet using WalletConnect login
- Web sign in with Rainbow wallet using Rainbow login
- Web sign in with Rainbow wallet using WalletConnect login
- Web sign in with Metamask wallet

Reviewers: rohan, atul, ginsu

Subscribers: