[lib][web][landing] Wagmi 1.0
Summary:
As part of working on [ENG-5152](https://linear.app/comm/issue/ENG-5152/siwe-broken-on-mobile), I updated RainbowKit and Wagmi in my local environment to see if it would resolve the issue. It eventually turned out that the issue did not have anything to do with the older RainbowKit and Wagmi versions, but I decided to diff up the upgrade anyways, in the interest of staying up-to-date on these package versions.

Depends on D9433

Test Plan:
1. In combination with other diffs in the stack, I tested the following on both my local environment, and in a staging environment for my keyserver on comm.domains:
  - Mobile sign in with Rainbow wallet using Rainbow login
  - Mobile sign in with Rainbow wallet using WalletConnect login
  - Web sign in with Rainbow wallet using Rainbow login
  - Web sign in with Rainbow wallet using WalletConnect login
  - Web sign in with Metamask wallet
2. Checked `node_modules/@rainbow-me/rainbowkit/dist/components/index.css` to confirm that the two CSS selector class names we override in `landing/siwe.css` still appeared in there
3. Confirm that the updated patchfile appears correctly

Reviewers: rohan, atul, ginsu

Subscribers: tomek, wyilio

Differential Revision: https://phab.comm.dev/D9434