[lib/web/native] Display community role for users in subchannels/sidebars
Summary: Currently, the members modal will display the correct role for community roots (i.e. 'Members', 'Admins'), but will only display 'Members' for subchannels/sidebars. This is because of the way that
`ThreadInfo` only contains the `Members` role for these threads.

The solution here is to introspect into the community's threadInfo for that specific thread, and pull the roles from there for each member. This will ensure that we have access to all of the community roles.

I noted in a code comment, but the special case is for `GENESIS`, where we don't readily have access to the community's members (and also roles are not really supported there), so in that case
we'll just read the current `ThreadInfo` for the role (probably 'Members' for subchannels/sidebars).

Resolves https://linear.app/comm/issue/ENG-5178/ashoat-shows-up-under-admins-column-but-with-members-pill

Test Plan: Please see the test videos below (also tested for custom roles)

Before:

After:

Reviewers: atul, ginsu, ashoat

Subscribers: