[keyserver] implement function to get account prekey and prekey signature
Summary:
Issue: [ENG-3944](https://linear.app/comm/issue/ENG-3944/update-keyserver-prekey-lifecycle-code-to-upload-to-identity-service#comment-8beb6586).

This is part of D8476 (breaking this for smaller diffs because it's complex).

Extracting code to separate function, will be used later in the stack so this reduces code duplication.

Depends on D9419

Test Plan:
Flow, this only moves some code around.

Reviewers: varun, marcin, ashoat