[keyserver] delete marking prekeys as published each time `olm` account is accessed
Summary:
Issue: [ENG-3944](https://linear.app/comm/issue/ENG-3944/update-keyserver-prekey-lifecycle-code-to-upload-to-identity-service#comment-8beb6586).

This is part of D8476 (breaking this for smaller diffs because it's complex).

Keys should be marked as published only on:
- [login](https://github.com/CommE2E/comm/blob/cd01c1732bd6534a0db2cb8d30bb0a30a48051f8/keyserver/src/user/login.js#L139)
- [register](https://github.com/CommE2E/comm/blob/cd01c1732bd6534a0db2cb8d30bb0a30a48051f8/keyserver/src/user/login.js#L158)
- [rotation](D9421)

Depends on D9421

Test Plan: N/A (but tested in previous diff that after actual publishing, keys are marked as published).

Reviewers: varun, marcin, ashoat

Reviewed By: marcin, ashoat

Subscribers: tomek, wyilio

Differential Revision: https://phab.comm.dev/D9422