[keyserver] script to publish prekeys to identity
Summary:
Issue: [ENG-3944](https://linear.app/comm/issue/ENG-3944/update-keyserver-prekey-lifecycle-code-to-upload-to-identity-service#comment-8beb6586).

Currently, prekeys on the prod keyserver are for sure marked as published, but yet published to Identity. To fix that uploading those via script.

Keys are already marked as published, so not marking those again to keep the old publishing date, only uploading to the Identity service.

When this diff is accepted I'll create a task for @ashoat to run this on production.

Depends on D9423

Test Plan:
1. `yarn script dist/scripts/publish-prekeys.js`.
2. Check for logs if `Prekeys published successfully`.
3. Check Identity logs.

Reviewers: varun, marcin, ashoat