[lib] Introduce a canBePinned boolean to message specs
Summary:
This diff introduces a `canBePinned` boolean to each of the message specs, isolating the 'source of truth' to the message specs. The next diff will have a function that will be able to read each of these, alongside unit tests and updating appropriate callsites.

Only composable message types can be pinned

Resolves https://linear.app/comm/issue/ENG-5213/introduce-a-canbepinned-boolean-to-message-specs

Depends on D9331

Test Plan: Ran `flow` for this diff, more thorough testing will be in the next diff

Reviewers: atul, ginsu

Reviewed By: atul

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9444