[web] update RelationshipPromptButton to be more reusable
Summary:
I want to use `RelationshipPromptButton` in user profiles. Right now there are some styles that are making it difficult to do that. In this diff I moved/changed some of the styles around + snuck in a quick memoization improvement to this component

Depends on D9431

Test Plan: Confirmed that `RelationshipPromptButton` looked and behaved the same

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9446