[web] introduce UserProfileMenu
Summary:
This diff introduces the `UserProfileMenu` component. This component handles showing the correct relationship action options based on what the current relationship status of the viewer and the profile is.

Here are the figma designs:

When user is a friend:
{F810089}

When viewer is blocked (couldn't find native designs but the concept is the same here):
{F810090}

When user is blocked by viewer:
{F810091}

When there is no relationship status between user/viewer, or there is a pending friend request in either direction:
{F810093}

Part of https://linear.app/comm/issue/ENG-5225/introduce-userprofilemenu

Depends on D9456

Test Plan:
Please see the screenshots below. Also confirmed that the actions worked as expected

When user is a friend:
{F810094}

When viewer is blocked:
{F810097}

When user is blocked by viewer:
{F810096}

When there is no relationship status between user/viewer, or there is a pending friend request in either direction:
{F810095}

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9457