[web] introduce usePushUserProfileModal hook
Summary:
This diff introduces `usePushUserProfileModal` hook which is going to be used to push the user profile modal. This logic will need to be repeated in every place we want the user profiles to be accessible, so I thought that instead of copy+pasting the logic we should preemptively factor out this logic into this helper hook.

Linear task: https://linear.app/comm/issue/ENG-5246/introduce-hook-to-push-user-profile-modal

Test Plan: Confirmed that I can push the user profile modal (the subsequent diffs will have the demo videos for each of the access points)

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9476