[web] introduce FullScreenViewModal
Summary:
To view a full screen avatar for user profiles we are going to want to reuse parts of the `MultimediaModal` component.

However, right now there are parts of `MultimediaModal` that are specific to viewing a multimedia message. We need to factor those parts out, so I created `FullScreenViewModal` which will be a generic component that can be used in both cases for image multimedia messages and user avatars.

Outlined below are the steps I will take in this stack (each point here will be it's own diff):
1. Introduce `FullScreenViewModal` file and move everything from the `ImageModal` file into here, and change what is absolutely necessary (the imports)
2. Lift the rendering of `mediaModalItem` out of `FullScreenViewModal`
3. Finish renaming + updating remaining variables/callbacks to make `FullScreenViewModal` completly generic
4. Introduce `UserProfileAvatarModal`

Linear Task: https://linear.app/comm/issue/ENG-5200/introduce-generic-fullscreenviewmodal

Depends on D9488

Test Plan: Confirmed that there were no regressions with the viewing a multimedia message in a full screen view

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9495