[web] factor out mediaModalItem from FullScreenModal
Summary:
In this diff, I factored out `mediaModalItem` from `FullScreenViewModal` and moved that logic into `MultimediaModal`. This diff also factors out the necessary styles, and the `dimensions` state which is used to help render `mediaModalItem`.

This is step 2 in the list below
Outlined below are the steps I will take in this stack (each point here will be it's own diff):
1. Introduce `FullScreenViewModal` file and move everything from the `ImageModal` file into here, and change what is absolutely necessary (the imports)
2. Lift the rendering of `mediaModalItem` out of `FullScreenViewModal`
3. Finish renaming + updating remaining variables/callbacks to make `FullScreenViewModal` completly generic
4. Introduce `UserProfileAvatarModal`

Part of this linear task: https://linear.app/comm/issue/ENG-5201/factor-out-viewing-image-multimedia-messages-specific-code-from

Depends on D9495

Test Plan: Confirmed that there were no regressions with viewing a multimedia message in a full screen view

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9496