[shared] helper functions to get/set version unsupported status
Summary:
we'll use the setter on the identity service but it might be useful for other services in the future, too.
we'll check for the specific unsupported version status on the client side.

Depends on D9473

Test Plan: tested the setter with kreya by sending a ping (implemented later in stack). will test is_unsupported_version when i handle the status on the client side.

Reviewers: bartek, michal

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9478