[lib] Remove `lastNavigatedTo/lastPruned` from `MessageStore.threads`
Summary:
All of the message pruning logic is now dependent on `lastNavigatedTo` and `lastPruned` from `ThreadActivityStore` instead of `MessageStore.threads`, so instances of `MessageStore.threads.`[lastNavigatedTo/lastPruned]` can safely be removed.

This diff just handles removal on the "JS side." We will separately handle removal on the C++ side.

NOTE: I originally intended to migrate values from `MessageStore.threads` to `ThreadActivityStore`, but decided against it. I will still introduce a migration to remove the extraneous `lastNavigatedTo` and `lastPruned` from `MessageStore.threads`, but figure it makes sense to start with ThreadActivityStore empty and be populated lazily instead of immediately filling with up with entries for every thread.

---

Depends on D9469

Test Plan: Putting this up now, but need to do more testing before landing so marking this as DRAFT. This should all be "dead code."

Reviewers: ashoat, tomek, ginsu, rohan, kamil

Reviewed By: ashoat, kamil

Subscribers: kamil, wyilio

Differential Revision: https://phab.comm.dev/D9475