[web] Add protoc-gen-js dependency
Summary:
This is an addon to `protoc` that is necessary in order to run the script in the next diff. It's not usually packaged as an NPM package - context [here](https://www.npmjs.com/package/protoc-gen-js).

The `protoc` in my local install has this addon already, which is why I didn't seem to need this change in my previous work in D7306. But the `protoc` in Nix doesn't seem to have this for some reason.

We could probably bring this in via Nix instead. I'm open to that if people prefer.

Test Plan:
Without this diff, the script in the following diff fails with the following error:

```
yarn run v1.22.19
$ ./scripts/codegen-identity-grpc.sh
protoc-gen-js: program not found or is not executable
Please specify a program using absolute path or make sure the program is available in your PATH system variable
--js_out: protoc-gen-js: Plugin failed with status code 1.
error Command failed with exit code 1.
info Visit https://yarnpkg.com/en/docs/cli/run for documentation about this command.
```

After introducing the `protoc-gen-js` dependency, the script no longer errors.

Reviewers: varun, atul

Subscribers: