[web] Introduce codegen-identity-grpc script
Summary:
In order to update our `grpc-web` codegen, we need to run `protoc` with some specific parameters, and then rename the files. I figured it would be good to put this into a Yarn script so we don't have to remember it / pull it from some diff's test plan every time.

Note that after running this script, some more work is necessary. There are a couple small modifications necessary to the `.cjs` files (update the imported filename and add `@generated` annotations), but the bulk of the work is in converting the TypeScript libraries to Flow.

Depends on D9490

Test Plan: Run the script

Reviewers: varun, atul

Reviewed By: varun

Subscribers: tomek, wyilio

Differential Revision: https://phab.comm.dev/D9491