[lib] Fix universal community permissions to prevent input bar and join chat button from appearing at the same time
Summary: Currently, there is a problem with the permissions we guarantee to all communities (`descendant_open_voiced`). This is actually not necessary, and @ashoat's suggestion fixed it:

> We should probably change this so that membership in the individual chat confers voiced. I think we can simply remove descendant_open_voiced from the community, and just leave it as a simple voiced. The permissions for each individual chat should confer voiced to members.

I should add a migration for this, it'll come as a follow up

Resolves https://linear.app/comm/issue/ENG-5180/join-chat-mode-showing-chat-input-field

Test Plan: Verified in an existing community (i.e. with the incorrect permission), the join chat button and the input field appears. In my new community created after removing this permission from being included, it's either one or the other.

Reviewers: atul, ginsu

Subscribers: