[web/keyserver] Separate web and keyserver versions
Summary:
[ENG-4766](https://linear.app/comm/issue/ENG-4766)

We need to separate web and keyserver for the future multi keyserver world. One of the things we need to change is versioning - they both need to have a separate versions.

Test Plan: `yarn flow-all`, `cargo test` in rust node addon. Checked if after changing the version the version in generated rust code also changes.

Reviewers: inka, kamil, atul

Reviewed By: atul

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9503