[keyserver] fix production check while uploading keys to Identity
Summary:
Fixing regression introduced in D9427.

This shouldn't affect anything, the only risk is that migration could make a silent fail on prod because the condition returns false always - but this is not what we want.

I discovered on my local setup that `process.env.NODE_ENV === 'production'` will not work because this env value is never set to production (this might be the fault of some missing `.env` files on my setup), but based on the code we only set `NODE_ENV` for dev, so changing this for safer solution.

Test Plan:
1. Terminate Identity
2. Test migration on prod (fail) / dev (only warning)

Reviewers: michal, marcin, ashoat

Reviewed By: ashoat

Subscribers: tomek, wyilio

Differential Revision: https://phab.comm.dev/D9505