[identity][docker] Add option to generate opaque keypair
Summary:

I'm aware that this functionality existed before, but was removed in D8580. Now I'm reintroducing it but making opt-in.
For commtest being run as a docker container that depends on identity, the easisest way is to generate the opaque server setup during image build.

Depends on D9507

Test Plan:

- `docker compose build identity-server` creates unchanged image
- `docker compose build --build-arg 'generate_keypair=true'` adds additional layer and generates `secrets/server_setup.txt`

Image layers can be checked using e.g. the `dive` tool: `dive commapp/identity-server`.

Reviewers: varun!, michal

Subscribers: