[CommCoreModule] Remove `last[NavigatedTo/Pruned]` from `MessageStoreThread` and `message_store_threads` table
Summary:
IMPORTANT: While I tested this migration a few times and read through things as carefully as I could, making changes to C++ codebase and SQLite store can be risky. Adding @ashoat and @kamil as blocking to take a look... would appreciate careful review to ensure everything is correct.

In this diff we:
1. Remove `lastPruned`/`lastNavigatedTo` from `MessageStoreThread` struct and remove usages on C++ side.
2. Introduce SQLite migration that drops `lastPruned` and `lastNavigatedTo` columns.

As of D9475, we've started ignoring what's persisted here in SQLite entirely... so this is effectively just dead code that we're cleaning up.

---

Depends on

D9475

Test Plan:
1. C++ codebase continues to build/run without issue.
2. `recreate_message_store_threads_table` migration succeeded and new `message_store_threads` table is as expected:
{F821730}

Before:
{F821724}

After:
{F821725}

Reviewers: ashoat, kamil, tomek, ginsu, rohan

Reviewed By: ashoat, kamil

Subscribers: wyilio, ashoat, kamil

Differential Revision: https://phab.comm.dev/D9517