Extract shareable notifications session creation logic to lib
Summary: Most of the logic implemented in `useInitialNotificationsEncryptedMessage` hook in `native` folder can be reused for olm session initialisation on web. That said the shareable part it should be moved to `lib` and only the platform specific part should be kept in `native`. Additionally this differential introduces necessary changes in native components (mostly boilerplate code).

Test Plan:
Using native app ensure that brand new olm session for notifs (`select * from olm_sessions;` in mariadb console) is created when:
1. Logging in.
2. Registering.
3. Registering with siwe.

Additionally repeat the second part of test plan for D9532 but for native app.

This will ensure that existing functionality relying on olm data on native isn't broken by this differential.

Reviewers: atul, michal, kamil

Reviewed By: atul

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9534