[keyserver/lib] Restrict the derivedMessages parameter from being passed to rawMessageInfoFromServerDBRow
Summary:
This should hopefully be the last step of [[ https://linear.app/comm/issue/ENG-4849/thread-from-robotext-about-pinned-message-causing-errors | ENG-4849 ]]. Now that we have a function `parseDerivedMessages` defined only in the message specs that will need to support derived messages (`SIDEBAR_SOURCE` and `TOGGLE_PIN`), we want to update the callsites of `rawMessageInfoFromServerDBRow` in `message-fetchers` to only pass in `derivedMessages` if necessary. This means for `rawMessageInfoFromServerDBRow` to receive `derivedMessages`, the message spec must have a `parseDerivedMessages` function in it.

Regarding the callsites, although we definitely don't need to pass in `derivedMessages` for the multimedia spec (I checked that it's not destructred), just to have it resistant to change I've only excluded it if `parseDerivedMessages` is not defined. For all other message specs, the same logic applies as well.

Resolves https://linear.app/comm/issue/ENG-5320/restrict-the-derivedmessages-parameter-from-being-passed-to

Depends on D9542

Test Plan: Ran `flow` and manually tested the web app to confirm that there are no regressions with this change

Reviewers: atul, ginsu

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9543