[web/native] Prevent text/multimedia from being pinned if they begin a sidebar
Summary: In the same way we handle reactions, we should prevent users from pinning messages (specifically text/multimedia) within a sidebar if they are the first message in the sidebar. We'll still allow other messages to be pinned in the sidebar.

Resolves https://linear.app/comm/issue/ENG-4847/not-possible-to-pin-the-source-messge

Depends on D9543

Test Plan: Confirmed that pinning messages in both a top-level thread and a sidebar still work as expected, and confirmed that the pin option doesn't appear for the first message in a sidebar on both `web` and `native`

Reviewers: atul, ginsu

Subscribers: